move errorBoundary to common
